Developer

+++ notification-service/src/main/java/org/example/service/TaskNotificationService.java
@@ -1,5 +1,6 @@
 package org.example.service;
 
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.example.dto.TaskNotificationDTO;
@@ -17,7 +18,7 @@ public class TaskNotificationService {
     private final TaskNotificationRepository taskNotificationRepository;
 
     public List<TaskNotification> getAllTaskNotificationsInRange(LocalDateTime start, LocalDateTime end) {
-        return taskNotificationRepository.findAllByDeadlineBeforeAndDeadlineAfter(start, end);
+        return taskNotificationRepository.findAllByDeadlineBeforeAndDeadlineAfter(end, start);
     }
 
     public void save(TaskNotification taskNotification) {
@@ -38,6 +39,11 @@ public class TaskNotificationService {
     public void delete(TaskNotificationDTO taskNotificationDTO) {
         TaskNotification taskNotification = taskNotificationRepository.
                 findByTaskIdAndTaskType(taskNotificationDTO.taskId(), taskNotificationDTO.taskType());
+
+        if (taskNotification == null) {
+            throw new EntityNotFoundException("TaskNotification not found for taskId: "
+                    + taskNotificationDTO.taskId() + " and taskType: " + taskNotificationDTO.taskType());
+        }
 
         taskNotificationRepository.delete(taskNotification);
     }
